Changed Teams router to a generic router Moved logging to msLogger
