+++ routers/tournamentDataRouter.js
@@ -3,7 +3,7 @@ const express = require('express')
 
 const router = express.Router()
 const MongoClient = require('mongodb').MongoClient
-const msLogger = require('@first-lego-league/ms-logger')
+const MsLogger = require('@first-lego-league/ms-logger').Logger
 
 const tournamentDataParser = require('../logic/tournamentDataParser')
 
@@ -21,35 +21,27 @@ router.post('/', (req, res) => {
   const data = tournamentDataParser.parse(req.body.tourData, req.body.delimiter)
   MongoClient.connect(process.env.MONGO).then(conn => {
     conn.db().collection('tables').insertMany(data.tables).then(() => {
-      // TODO: use ms-logger
-      console.log('all good')
+      MsLogger.info('Data saved successfully to collection tables')
     }).catch(err => {
-      // TODO: use ms-logger
-      console.log(err)
+      MsLogger.error('Something went wrong while saving data \n' + err)
     })
 
     conn.db().collection('teams').insertMany(data.teams).then(() => {
-      // TODO: use ms-logger
-      console.log('all good')
+      MsLogger.info('Data saved successfully to collection teams')
     }).catch(err => {
-      // TODO: use ms-logger
-      console.log(err)
+      MsLogger.error('Something went wrong while saving data \n' + err)
     })
 
     conn.db().collection('practice-matches').insertMany(data.practiceMatches).then(() => {
-      // TODO: use ms-logger
-      console.log('all good')
+      MsLogger.info('Data saved successfully to collection practice-matches')
     }).catch(err => {
-      // TODO: use ms-logger
-      console.log(err)
+      MsLogger.error('Something went wrong while saving data \n' + err)
     })
 
     conn.db().collection('ranking-matches').insertMany(data.rankingMatches).then(() => {
-      // TODO: use ms-logger
-      console.log('all good')
+      MsLogger.info('Data saved successfully to collection ranking-matches')
     }).catch(err => {
-      // TODO: use ms-logger
-      console.log(err)
+      MsLogger.error('Something went wrong while saving data \n' + err)
     })
 
     res.status(201)
